Extract file extension icon jsx into separate component
